Create meta JS project .
We have to create a variable that can hold a number of NFT's , after which we have to create an object inside our mintNFT function which will hold the metadata of NFT's. Then the list NFT's will print all of our NFT's metadata to console . Afterall we have to return the number of NFT's we have created.